+++ packages/connect/src/index.js
@@ -6,5 +6,5 @@ export {
   raw,
 } from "@nx-js/observer-util";
 export { createStore, getSnapshot } from "./create-store";
-export { connect as default, Provider } from "./connect";
+export { connect as default, Provider, useConnect } from "./connect";
 export { batch } from "@risingstack/react-easy-state";