+++ packages/frontity/src/steps/index.ts
@@ -116,6 +116,22 @@ export const createPackageJson = async (
   await writeFile(filePath, fileData);
 };
 
+// This function create a `README.md` file.
+export const createReadme = async (
+  name: string,
+  path: string
+): Promise<void> => {
+  const fileTemplate = await readFile(
+    resolvePath(__dirname, "../../templates/README.md"),
+    {
+      encoding: "utf8",
+    }
+  );
+  const filePath = resolvePath(path, "README.md");
+  const fileData = fileTemplate.replace(/\$name\$/g, name);
+  await writeFile(filePath, fileData);
+};
+
 // This function creates a `frontity.settings` file.
 export const createFrontitySettings = async (
   extension: string,
@@ -127,7 +143,7 @@ export const createFrontitySettings = async (
     name,
     state: {
       frontity: {
-        url: "https://test.frontity.io",
+        url: "https://test.frontity.org",
         title: "Test Frontity Blog",
         description: "WordPress installation for Frontity development",
       },
@@ -155,7 +171,7 @@ export const createFrontitySettings = async (
         name: "@frontity/wp-source",
         state: {
           source: {
-            api: "https://test.frontity.io/wp-json",
+            api: "https://test.frontity.org/wp-json",
           },
         },
       },