+++ packages/hooks/index.ts
@@ -1,0 +1,1 @@
+export { default as useInView } from "./use-in-view";