+++ packages/mars-theme/src/components/link.js
@@ -1,41 +1,39 @@
 import React from "react";
-import { connect } from "frontity";
+import { connect, useConnect } from "frontity";
+import Link from "@frontity/components/link";
 
-const Link = ({
-  state,
-  actions,
-  link,
-  className,
-  children,
-  "aria-current": ariaCurrent,
-}) => {
-  const onClick = (event) => {
-    // Do nothing if it's an external link
-    if (link.startsWith("http")) return;
+/**
+ * The MarsLink component, which is a wrapper on top of the {@link Link}
+ * component.
+ *
+ * @param props - It accepts the same props than the {@link Link} component.
+ *
+ * @example
+ * ```js
+ * <MarsLink link="/some-post">
+ *   <div>Some Post</div>
+ * </MarsLink>
+ * ```
+ *
+ * @returns A {@link Link} component, which returns an HTML anchor element.
+ */
+const MarsLink = ({ children, ...props }) => {
+  const { state, actions } = useConnect();
 
-    event.preventDefault();
-    // Set the router to the new url.
-    actions.router.set(link);
-
-    // Scroll the page to the top
-    window.scrollTo(0, 0);
-
-    // if the menu modal is open, close it so it doesn't block rendering
+  /**
+   * A handler that closes the mobile menu when a link is clicked.
+   */
+  const onClick = () => {
     if (state.theme.isMobileMenuOpen) {
       actions.theme.closeMobileMenu();
     }
   };
 
   return (
-    <a
-      href={link}
-      onClick={onClick}
-      className={className}
-      aria-current={ariaCurrent}
-    >
+    <Link {...props} onClick={onClick}>
       {children}
-    </a>
+    </Link>
   );
 };
 
-export default connect(Link);
+export default connect(MarsLink, { injectProps: false });