+++ packages/wp-source/src/libraries/handlers/__tests__/category.tests.ts
@@ -14,7 +14,7 @@ let store: InitializedStore<WpSource>;
 let api: jest.Mocked<Api>;
 beforeEach(() => {
   store = createStore(clone(wpSource()));
-  store.state.source.api = "https://test.frontity.io/wp-json";
+  store.state.source.api = "https://test.frontity.org/wp-json";
   store.actions.source.init();
   api = store.libraries.source.api as jest.Mocked<Api>;
 });
@@ -149,6 +149,28 @@ describe("category", () => {
       );
     // Fetch entities
     await store.actions.source.fetch("/category/cat-1/?some=param");
+    expect(store.state.source).toMatchSnapshot();
+  });
+
+  test("Unknown URL should return a 404 even if it's substring matches a path", async () => {
+    api.get = jest.fn((_) =>
+      Promise.resolve(
+        mockResponse(cat1Posts, {
+          "X-WP-Total": "5",
+          "X-WP-TotalPages": "2",
+        })
+      )
+    );
+
+    await store.actions.source.fetch("/category/undefined/cat-1/");
+
+    expect(
+      (store.state.source.data as any)["/category/undefined/cat-1/"]
+        .errorStatusText
+    ).toBe(
+      "You have tried to access content at route: /category/undefined/cat-1/ but it does not exist"
+    );
+
     expect(store.state.source).toMatchSnapshot();
   });
 });