+++ packages/wp-source/src/libraries/handlers/postType.ts
@@ -2,11 +2,51 @@ import { Handler } from "../../../types";
 import capitalize from "./utils/capitalize";
 import { ServerError } from "@frontity/source";
 
+/**
+ * The parameters for {@link postTypeHandler}.
+ */
+interface PostTypeHandlerParams {
+  /**
+   * The list of [WP REST API endpoints](https://developer.wordpress.org/rest-api/reference/)
+   * from which the generated handler is going to fetch the data.
+   */
+  endpoints: string[];
+}
+
+/**
+ * A {@link Handler} function generator for WordPress Post Types.
+ *
+ * This function will generate a handler function for specific
+ * [WP REST API endpoints](https://developer.wordpress.org/rest-api/reference/)
+ * from which the data is going to be fetched. The generated handler will fetch
+ * data from all specified endpoints.
+ *
+ * @param options - Options for the handler generator: {@link PostTypeHandlerParams}.
+ *
+ * @example
+ * ```js
+ *   const postTypeHandlerFunc = postTypeHandler({ endpoints: ['post']});
+ *   libraries.source.handlers.push({
+ *     name: "post type",
+ *     priority: 30,
+ *     pattern: "/(.*)?/:slug",
+ *     func: postTypeHandlerFunc,
+ *   })
+ * ```
+ *
+ * @returns An async "handler" function that can be passed as an argument to the handler object.
+ * This function will be invoked by the frontity framework when calling `source.fetch()` for
+ * a specific entity.
+ */
 const postTypeHandler = ({
   endpoints,
-}: {
-  endpoints: string[];
-}): Handler => async ({ link, params, state, libraries }) => {
+}: PostTypeHandlerParams): Handler => async ({
+  link,
+  params,
+  state,
+  libraries,
+  force,
+}) => {
   // 1. search id in state or get the entity from WP REST API
   const { route, query } = libraries.source.parse(link);
   if (!state.source.get(route).id) {
@@ -19,19 +59,38 @@ const postTypeHandler = ({
 
     // 1.2 iterate over finalEndpoints array
     let isHandled = false;
+    let isMismatched = false;
     for (const endpoint of finalEndpoints) {
       const response = await libraries.source.api.get({
         endpoint,
         params: { slug, _embed: true, ...state.source.params },
       });
 
-      const populated = await libraries.source.populate({ response, state });
+      const populated = await libraries.source.populate({
+        response,
+        state,
+        force,
+      });
 
       // exit loop if this endpoint returns an entity!
       if (populated.length > 0) {
-        isHandled = true;
-        break;
+        // We have to check if the link property in the data that we
+        // populated is the same as the current route.
+        if (populated[0].link === route) {
+          isHandled = true;
+          isMismatched = false;
+          break;
+        } else {
+          isMismatched = true;
+        }
       }
+    }
+
+    if (isMismatched) {
+      throw new ServerError(
+        `You have tried to access content at route: ${route} but it does not exist`,
+        404
+      );
     }
 
     // 1.3 if no entity has found, throw an error