+++ packages/core/CHANGELOG.md
@@ -1,5 +1,77 @@
 # Change Log
 
+## 1.7.3
+
+### Patch Changes
+
+- [`a0fc05cb`](https://github.com/frontity/frontity/commit/a0fc05cb8d51e4e101994b1b35410d5c2fd16e55) [#525](https://github.com/frontity/frontity/pull/525) Thanks [@luisherranz](https://github.com/luisherranz)! - Fix for package names that start with a number, like `@123/package`.
+
+## 1.7.2
+
+### Patch Changes
+
+- [`322d22ec`](https://github.com/frontity/frontity/commit/322d22ecb825d510296243736a79e4208023477f) [#501](https://github.com/frontity/frontity/pull/501) Thanks [@michalczaplinski](https://github.com/michalczaplinski)! - Add the tsNode.register for `dev` and `build` scripts which were previously in `frontity`
+
+- Updated dependencies [[`159e02ca`](https://github.com/frontity/frontity/commit/159e02ca080ec9f7004c90276621d1a2708192ce)]:
+  - @frontity/connect@1.1.3
+
+## 1.7.1
+
+### Patch Changes
+
+- [`2c9ddab1`](https://github.com/frontity/frontity/commit/2c9ddab1f8962c1860fe206f77d8872808c2c3ef) [#463](https://github.com/frontity/frontity/pull/463) Thanks [@juanmaguitar](https://github.com/juanmaguitar)! - README.md with full info of the package.
+
+* [`aade774f`](https://github.com/frontity/frontity/commit/aade774fba18e8790153e49365790a1c7826a9a5) [#467](https://github.com/frontity/frontity/pull/467) Thanks [@luisherranz](https://github.com/luisherranz)! - Do not allow the new `state.frontity.debug` flag to be `true` in production.
+
+* Updated dependencies [[`ba11bca8`](https://github.com/frontity/frontity/commit/ba11bca879d86493181b35e358ded8c583496641), [`1440cfe7`](https://github.com/frontity/frontity/commit/1440cfe77f1a9154562dff23a97185c77bebb59c), [`b6ada1cc`](https://github.com/frontity/frontity/commit/b6ada1cc7075d32a9ef82d525e251e1554baaba3)]:
+  - babel-plugin-frontity@1.0.2
+  - @frontity/connect@1.1.2
+  - @frontity/file-settings@1.1.6
+
+## 1.7.0
+
+### Minor Changes
+
+- [`4212f46a`](https://github.com/frontity/frontity/commit/4212f46a9a655b2c91940135a22f3f0b69715119) [#422](https://github.com/frontity/frontity/pull/422) Thanks [@Koli14](https://github.com/Koli14)! - Optionally, use the content of a robots.txt file if it exists in the root project folder.
+
+### Patch Changes
+
+- [`f13f72bb`](https://github.com/frontity/frontity/commit/f13f72bb19b7d1f379ef21e349e7d6b047ec5098) [#456](https://github.com/frontity/frontity/pull/456) Thanks [@DAreRodz](https://github.com/DAreRodz)! - Replace `lodash-es` by `lodash` library and remove the `lodash` alias to avoid problems with third-party libraries.
+
+  Also, adds `babel-plugin-lodash` to still get tree shaking when importing named exports from `lodash`.
+
+* [`5cb29ab6`](https://github.com/frontity/frontity/commit/5cb29ab63ab31872a4d853e5e2fdbdabca974c9f) [#430](https://github.com/frontity/frontity/pull/430) Thanks [@luisherranz](https://github.com/luisherranz)! - Remove React `Fills` package export. This was meant to be used for the [Slot and Fill](https://community.frontity.org/t/slot-and-fill/895) pattern, but we are going to finally use `state.fills` to add the configuration for the fills and `libraries.fills` to expose the React components.
+
+- [`5ada3cb0`](https://github.com/frontity/frontity/commit/5ada3cb0e44dc869291d5e9fdea67a1ef6bcced8) [#374](https://github.com/frontity/frontity/pull/374) Thanks [@michalczaplinski](https://github.com/michalczaplinski)! - Try to re-use an already open browser tab when running `frontity dev` on MacOS.
+
+- Updated dependencies [[`5cb29ab6`](https://github.com/frontity/frontity/commit/5cb29ab63ab31872a4d853e5e2fdbdabca974c9f)]:
+  - @frontity/connect@1.1.1
+
+## 1.6.1
+
+### Patch Changes
+
+- [`b4b25e43`](https://github.com/frontity/frontity/commit/b4b25e434e5e7c2a96e1fce8c11b0e0dac228295) [#446](https://github.com/frontity/frontity/pull/446) Thanks [@DAreRodz](https://github.com/DAreRodz)! - Fix --publicPath option when its value is a URL with domain specified.
+
+## 1.6.0
+
+### Minor Changes
+
+- [`661ac083`](https://github.com/frontity/frontity/commit/661ac08316f44172166e79b05b47f0c15a837a9a) [#420](https://github.com/frontity/frontity/pull/420) Thanks [@DAreRodz](https://github.com/DAreRodz)! - Add a `--publicPath` option to `build` and `dev` commands and pass it to webpack builds.
+
+### Patch Changes
+
+- Updated dependencies [[`996865a2`](https://github.com/frontity/frontity/commit/996865a27690d5b89d2ef110f5b1bf3fb91da6f5)]:
+  - @frontity/connect@1.1.0
+
+## 1.5.3
+
+### Patch Changes
+
+- [`502bc144`](https://github.com/frontity/frontity/commit/502bc144158a1b64971c738dcf93306ad5b61a8f) [#406](https://github.com/frontity/frontity/pull/406) Thanks [@luisherranz](https://github.com/luisherranz)! - Don't embed fonts so big because they have an impact on LightHouse performance score. I have reduced the limit from 25Kbs to 8Kbs, which seems to be the recommendation in Webpack.
+
+* [`7c2a99f9`](https://github.com/frontity/frontity/commit/7c2a99f9be43ad965f4d4b00c81145f21230f63f) [#387](https://github.com/frontity/frontity/pull/387) Thanks [@iamuchejude](https://github.com/iamuchejude)! - Fix yarn symlink bug
+
 ## 1.5.2
 
 ### Patch Changes