+++ packages/google-analytics/CHANGELOG.md
@@ -1,5 +1,29 @@
 # @frontity/google-analytics
 
+## 1.2.0
+
+### Minor Changes
+
+- [`5ece8821`](https://github.com/frontity/frontity/commit/5ece88219d354952baadd5649091b09dd5288b49) [#472](https://github.com/frontity/frontity/pull/472) Thanks [@DAreRodz](https://github.com/DAreRodz)! - Release the first stable version of `@frontity/google-analytics` package ([Feature Discussion](https://community.frontity.org/t/google-analytics-package/1084)).
+
+### Patch Changes
+
+- Updated dependencies [[`322d22ec`](https://github.com/frontity/frontity/commit/322d22ecb825d510296243736a79e4208023477f), [`f5bf7b1c`](https://github.com/frontity/frontity/commit/f5bf7b1cee2850445fe5304e1b39e20e786e9377), [`2fb73fc7`](https://github.com/frontity/frontity/commit/2fb73fc798653803a21f9e9bd7f21355f7675e55)]:
+  - frontity@1.10.0
+  - @frontity/analytics@1.2.0
+
+## 1.1.0
+
+### Minor Changes
+
+- [`72c33fbd`](https://github.com/frontity/frontity/commit/72c33fbde5d60de33e7f5c25f081ffd458d15f63) [#451](https://github.com/frontity/frontity/pull/451) Thanks [@juanmaguitar](https://github.com/juanmaguitar)! - README.md updated or added
+
+### Patch Changes
+
+- Updated dependencies [[`24a20ee1`](https://github.com/frontity/frontity/commit/24a20ee15e65d56f88daac4dd49372072bdd10c6), [`6900916a`](https://github.com/frontity/frontity/commit/6900916ace309d3cc55b9c732124df5d3db96838), [`72c33fbd`](https://github.com/frontity/frontity/commit/72c33fbde5d60de33e7f5c25f081ffd458d15f63)]:
+  - @frontity/analytics@1.1.0
+  - frontity@1.8.0
+
 ## 1.0.0
 
 ### Major Changes