+++ packages/hooks/CHANGELOG.md
@@ -1,5 +1,28 @@
 # Change Log
 
+## 2.0.2
+
+### Patch Changes
+
+- [`fc1220b5`](https://github.com/frontity/frontity/commit/fc1220b5eb2ae0af933b76cc9fb1a62b83fc7075) [#484](https://github.com/frontity/frontity/pull/484) Thanks [@juanmaguitar](https://github.com/juanmaguitar)! - README.md created.
+
+## 2.0.1
+
+### Patch Changes
+
+- [`5cb29ab6`](https://github.com/frontity/frontity/commit/5cb29ab63ab31872a4d853e5e2fdbdabca974c9f) [#430](https://github.com/frontity/frontity/pull/430) Thanks [@luisherranz](https://github.com/luisherranz)! - Mark the package as "side-effect free" to allow [tree shaking in Webpack](https://webpack.js.org/guides/tree-shaking/).
+
+## 2.0.0
+
+### Major Changes
+
+- [`8a0bb8f0`](https://github.com/frontity/frontity/commit/8a0bb8f03ad70ac6b92be14c2c95dcad2b3ccf75) [#413](https://github.com/frontity/frontity/pull/413) Thanks [@orballo](https://github.com/orballo)! - Reimplementation of `useInView` hook as a wrapper for `react-intersection-observer` that checks if `IntersectionObserver` is supported by the browser.
+
+### Patch Changes
+
+- Updated dependencies [[`661ac083`](https://github.com/frontity/frontity/commit/661ac08316f44172166e79b05b47f0c15a837a9a), [`75343be3`](https://github.com/frontity/frontity/commit/75343be3154dda5b587a3065b225161af96f0488)]:
+  - frontity@1.7.0
+
 ## 1.2.0
 
 ### Minor Changes